restructure repo for pip packaging
