+++ examples/mnist/spark/mnist_dist.py
@@ -13,7 +13,7 @@ def print_log(worker_num, arg):
   print("{0}: {1}".format(worker_num, arg))
 
 def map_fun(args, ctx):
-  from com.yahoo.ml.tf import TFNode
+  from tensorflowonspark import TFNode
   from datetime import datetime
   import math
   import numpy
