+++ examples/mnist/streaming/mnist_spark.py
@@ -19,7 +19,7 @@ import threading
 import time
 from datetime import datetime
 
-from com.yahoo.ml.tf import TFCluster
+from tensorflowonspark import TFCluster
 import mnist_dist
 
 sc = SparkContext(conf=SparkConf().setAppName("mnist_streaming"))
