+++ tensorflowonspark/TFManager.py
